use cleaned name for pluralized resources in cfg
Cleans up the `pkg/model:GetOpTypeAndResourceNameFromOpID()` function so
that resources with pluralized singular names can be referred to with
the "cleaned" Camel-cased name in the generator.yaml file. In addition,
corrects an error in said function where the operation returning a
singular pluralized resource -- e.g. DescribeDhcpOptions -- was
incorrectly being returned as OpTypeList instead of OpTypeGet.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>