Fix type of tolerations in helm chart (#267)
At the moment, the helm chart treats the deployment's tolerations as an object (similarly to nodeSelector). In the Deployment manifest however tolerations are an array (in contrast to nodeSelector). This PR makes the deployment's tolerations of type array instead of object.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.