Added missing pullSecrets & pullPolicy (#293)
Signed-off-by: Nick Fisher <nxf5025@gmail.com>

Description of changes: The helm chart's values.yaml files have the ability to set `pullPolicy` & `pullSecrets`, but the deployment.yaml does not contain the proper blocks to render the settings. This PR adds both of those.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.