Add `Region()` to `ResourceMetadata` struct and `AWSResourceIdentifiers` interface
Currently when a resource is created without setting a region in the
annotations, the controller uses either the namespace annotation or the
controller region flag. This causes the controller to miss manage the
resource if they are restart with a different region, or if run a second
controller using a different region flag.

This patch modifies the code-generator to generate a `Region` function
so that a resource can also satisfy the `Identifiers` interface. This
patch also extends the `setStatusDefaults` so that it correctly
populates the status.ResourceMetadata.Region with the aws region in
which it is created/managed

This patch will also require some changes to the runtime repository.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.