adding operator-sdk to the code-generator /bin folder (#305)
Issue #, if available:
N/A
Description of changes:
Since there is a change in operator-sdk that isn't compatible with the way ACK `deployment.yaml`'s are built, it's best if we install and use a specific version in the `/bin` folder of the `code-generator`. This way no matter what version of `operator-sdk` a developer has installed, we can control what versions are used for local development and release of this project. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.

Signed-off-by: Adam D. Cornett <adc@redhat.com>