Generate `Region()` method for `resourceIdentifiers` (#309)
Follow up to https://github.com/aws-controllers-k8s/runtime/pull/79

This patch modifies the code-generator to generate a `Region` function
so that a resource can also satisfy the `Identifiers` interface. This
the patch also extends the `setStatusDefaults` so that it correctly
populates the `status.ResourceMetadata.Region` with the aws region in
which it is created/managed

Description of changes:
- Add `Region()` method to resourceIdentifiers template.
- Modifies `setStatusDefaults` to  be able to set the a default region.
- Update `runtime_test.go`

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.