add real versions to controller main.go (#321)
I'd left a TODO in the main.go template for controllers to fill in their
GitVersion, GitCommit and BuildDate information properly. Just now
getting around to this...

Issue: aws-controllers-k8s/community#1171

Signed-off-by: Jay Pipes <jaypipes@gmail.com>

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.