Add support for references within lists of structs (#332)
Issue #, if available: https://github.com/aws-controllers-k8s/community/issues/1256

Description of changes:
This pull request adds support for references that occur within lists of structs. It updates the generated `references.go` files in each controller to now iterate through the structs to check for the existence of references and again for resolving their individual references. Rather than creating a new slice of structs with resolved references, it instead resolves the references in place - updating the base field value and leaving the reference untouched. 

This pull request has been tested against the APIGatewayv2 and EC2 resource reference tests, as well as manually against the EC2 changes that inspired the original issue.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.