Use metadata as source of truth for latest version (#348)
Fixes https://github.com/aws-controllers-k8s/community/issues/1361

Description of changes:
Code-generator currently compares the subdirectories in the API output directory (`apis/`) to determine the latest K8s version of the CRDs. The information about which API versions should be generated, and their status (`available`, `deprecated`, etc.), already exist in the `metadata.yaml` file. This pull request updates the source of truth for the code-generator to be the `metadata.yaml` file.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.