Return a terminal error when `sdkUpdate` is not implemented
Currently the generated sdk code throws a simple error when `sdkUpdate`
is not implemented. This is only observable in the controller logs and
users are not able to read this information from their resource's
status/conditions.

This patch forces `sdkUpdate` to return a terminal error instead of a
normal error.

Fixes https://github.com/aws-controllers-k8s/community/issues/1521

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>