Fix generation of attribute-based APIs (#376)
The code generation for attribute-based APIs (SNS, SQS, etc) had not been kept up to date for 2+ years, meaning that the SNS and SQS controllers could not be regenerated to bring in updated ACK runtimes or aws-sdk-go releases.

This commit gets the attribute-based API code generation back into shape. Most of the changes in here were around the code generator being able to find either ResourceConfig or FieldConfig objects using case-insensitive matching, which turned out to be the main cause for code generator failures when re-generating the SNS controller.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.