Add a configuration for `additional_columns` at the Resource level (#378)
Issue #, if available:  [1281](https://github.com/aws-controllers-k8s/community/issues/1281)

Description of changes:

We want to have the option to include additional printer columns in the output of `kubectl`.  This change allow the author to specify an arbitrary number of those per Resource, for each specifying the name, type, and JSON path of the field to produce.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.