Lint all scripts with shellcheck (#392)
Description of changes:
Used the output of the `shellcheck` tool to make style changes to all of the shell scripts within the repository. Added a new `make` target for `lint-shell` which lints all of the scripts and excludes SC1091, which fails because we `source` other scripts using path variables (a known limitation of `shellcheck`).

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.