Revert "Correct the generated code for nested parent field reference." (#398)
Fixes https://github.com/aws-controllers-k8s/community/issues/1653
Reverts aws-controllers-k8s/code-generator#369

This pull request introduced a change to the way lists of references were accessed, but incorrectly falls back on a conditional that will treat a list of references as a struct. This breaks the `ec2` and `rds` controllers.