feat: Allow manual setting of AWS client struct and interface names
While developing the ACK pipes-controller, we encountered an issue where
the generated code didn't compile mainly because the code-generator
assumed that the AWS SDK client/interface/import-paths always followed
the same naming convetions. However, we discovered that pipes was the
exception as they renamed the model, import path and client
struct/interface.

This patch introduces a new feature that enables users to manually set
the client structg and interface names int he `generator.yaml` file.
This update allows for greater flexibvility in the AWS client naming
within pipes-controller and possibly other controllers in the future.

In this patch, we've also introduced a breaking change by renaming
`model_name` to `sdk_names.model`. Mianly to keep all the SDK related
naming convetions in the same place.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>