support field custom documentation strings (#394)
Adds support for extending the documentation strings for fields and
setting docstrings for custom fields.

A new `FieldConfig.Documentation` configuration option is now supported.
When present, this string *extends* any documentation provided for the
existing field by the doc-2.json files. If no such documentation is
provided by the existing field OR if the field is a custom field, the
`FieldConfig.Documentation` configuration option is used for the field's
docstring in the Go type files.

Example usage for an existing nested field (from a hypothetical ec2
generator.yaml file). The
`LaunchTemplateData.HibernationOptions.Configured` field has a docstring
already that contains the following:

```
// If you set this parameter to true, the instance is enabled for hibernation.
//
// Default: false
```

Setting the `generator.yaml` file to the following:

```yaml
resources:
  LaunchTemplate:
    fields:
      LaunchTemplateData.HibernationOptions.Configured:
        documentation: XXX extended docs XXX
```

would cause the field's Documentation string to look like this:

```
// If you set this parameter to true, the instance is enabled for hibernation.
//
// Default: false
//
// XXX extended docs XXX",
```

Issue aws-controllers-k8s/community#1223

Signed-off-by: Jay Pipes <jaypipes@gmail.com>