Set Resource when GetAttribute is present (#439)
This change basically the SetResource function to continue when the `GetAttribute` operation is present. If the `ReadOne` operation is present then that is preferred.

Issue https://github.com/aws-controllers-k8s/community/issues/1778

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.