Update template to support watch namespace (#440)
As the controller runtime supports watching namespaces, we would need to add a support to watch namespaces other than where the controller is deployed. This is a common use case for multi-tenant K8s environments. This PR introduces a `watchNamespace` flag which lets user define the namespaces to be watched by the controller. If left blank, it would take the release namespace.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.