code-gen: needs to support custom find_operation
Issue aws-controllers-k8s/community#1827

Description of changes:
This change allow to implement custom find function. The name of the function could be written in
generator.yaml as 'find_operation: <name-of-function>' similar to update_operation. And generated
code for sdkFind() would call this custom find function. The find function itself could be written
in hooks.go.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.