Fix creation of multiple alarms (#4)
Issue #, if available: n/a

Description of changes: I discovered a bug in the controller's sdkFind logic that caused it to fail to create alarms in many cases. Basically, it was using `DescribeAlarms` with no alarm name input, so would get back a list of all alarms. If any alarms existed in the account already, it would determine the new alarm already existed and not create it.

This fix adds a post_build_request hook to read_many so it will get back a list of just the one alarm it's trying to find.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.