fix: ensure boolean flags are only passed when corresponding Helm chart value is true
In this commit, we address the issue of passing boolean flags in a more
solid way. We've observed an odd obehaviour when a user attempts to
disabled a feature (e.g `leaderElection/developmentLogging`), which
causes the flag to be enabled instead.

The main root cause of this issue was, using environment vairables with
the `pflag` library, combined with the fact that `spf13/cobra` library
conciders passed boolean flags with no values (no `=<value>`) as `true`

The fix consists of
1. Avoid using environment variables for boolean flags.
2. Updating the helm template to only pass boolean flags when their
   respctive helm chat values are present.

Signed-off-by: Amine <hilalyamine@gmail.com>