Support of injecting fields in existing structures: CustomNestedFields (#462)
**Issue #**
This issue arises from the code changes in lambda-controller for[ PR #88](https://github.com/aws-controllers-k8s/lambda-controller/pull/88)

**Summary**
Support of injecting fields in existing structures: CustomNestedFields

**Description**
The code in this PR helps in accepting a custom nestedField defined using `type` as a part of its parentField. In the issue mentioned above when we inject a new field `S3SHA256` into `Code` struct, instead of nesting it under `Code` the generator created a separate field named `CodeS3SHA256`. 

The new field `S3SHA256` is defined like this in `generator.yaml` file:
```
Function:
    fields:
      Code.S3SHA256:
        type: string
```

After applying the changes in this PR, the field `S3SHA256` is properly injected as a part of `Code` struct.

```
type FunctionCode struct {
	ImageURI *string `json:"imageURI,omitempty"`
	S3Bucket *string `json:"s3Bucket,omitempty"`
	S3BucketRef     *ackv1alpha1.AWSResourceReferenceWrapper `json:"s3BucketRef,omitempty"`
	S3Key           *string                                  `json:"s3Key,omitempty"`
	S3ObjectVersion *string                                  `json:"s3ObjectVersion,omitempty"`
	S3SHA256        *string                                  `json:"s3SHA256,omitempty"`
	ZipFile         []byte                                   `json:"zipFile,omitempty"`
} 
```

This PR contains code implementation to inject Nested fields into the Spec or Status struct and has respective model-test to validate the changes.

**Acknowledgment**
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.