fix(omitUnchangedFields): only generate delta comparison check for spec fields (#476)
Skips the code generation of delta comparison check triggered by the use of the `omit_unchanged_fields:` field, if the field in question comes from the resource's `Status`. Such fields should always be included in the request, even with negative delta.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.