Correctly handle watchNamespace flag parsing errors
Previously, we shipped a feature (d3b23cd) where multi-namespace watch
mode was supported. It turns out that the logic was completely flipped.
We need to assign namespace when the error is NIL and not vice versa. If
the error is NOT NIL, we need to block the controller initiation and
throw an error.

How did we catch this? Nothing special, just reviewing the generated
code and asking who on earth wrote that.

This indeed urges us to do proper CARM/multi-namespace/cross-region
testing and write some e2e tests that we can catch very early on. When
code-generator PRs are drafted.

The current infra doesn't allow us to that, so we'll hack something
around next week.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>