Experimental configuration to rename the controller name
Since the early days of ACK, we've always relied on the service ID to
name the controller. The way names, go modules, import paths, helm
charts are named relied on this serviceID. Which resulted in creating
some controllers named in an odd way, like
`prometheusservice-controller` instead of `amp-controller`,
`acmpca-controller` instead of `pca-controller`. This patch introduces
an experimental configuration that instructs the code generator to
override the serviceID when it comes to import paths/helm charts names
etc...

We will test this feature out with documentdb controller.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>