Default watchNamespace to release namespace
This patch default the `watchNamespace` to the release namespace if not
provided.

When a helm chart is installed users can set two parameters to define
the isntallation mode: `installScope`, `watchNamespace`. The latter is
only actionable when `installScope` is set to `namespace`. This patch
defaults the `watchNamespace` to the helm release namespace if the user
doesn't provide an This patch defaults the `watchNamespace` to the helm
release namespace if the user doesn't provide any,y,

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>