Default watchNamespace to release namespace (#501)
Issue https://github.com/aws-controllers-k8s/community/issues/2006

This patch default the `watchNamespace` to the release namespace if not
provided.

When a helm chart is installed users can set two parameters to define
the isntallation mode: `installScope`, `watchNamespace`. The latter is
only actionable when `installScope` is set to `namespace`. This patch
defaults the `watchNamespace` to the helm release namespace if the user
doesn't provide any.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.