Extend `getAttributeFromPath` search in status fields as well.
Prior to this patch, `getAttributeFromPath` searched for paths in spec
fields only. And acted as a no-op when it comes to status fields. This
patch extends the search to status fields as well, allowing us to
configure nested fields in the status as well.

This is gonna be used to configure ECS Cluster `Status.Attachments.Type`
go tags (set `type` instead of `type_`)

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>