Fix comparison type mismatch in deployment template
This commit addresses the `incompatible types for comparison`
error ncountered during the deployment process (using helm).
The issue was caused by the incorrect comparison of a integer against a
float

This patch fixesd the issues describe above by casting the
`reconcile.defaultResyncPeriod` value to an integerer before performing
the comparison.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>