feat(fieldpath): Handle case differences in structure member names
This patch updates the `memberShapeRef` function in the `fieldpath`
package to handle case differences between struct member names in the
ACK model and the AWS SDK Model.

Previously, the function assumed that member names in the ACK model and
AWS SDK model were identical. However, this is not always the case, as
some field names may have different casing (e.g `AWSVPCConfiguration` in
the ACK model and `AwsVpcConfiguration` in the AWS SDK model).

To address this issue, the function now leverages `strings.EqualFold` to
compare the member names case isensitively.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>