feature: Generate Code for resource adoption by annotation
These changes introduce a new generated function in all controllers
that attempts to populate the resource spec/status with fields
defined by the user. This change will require for developers to
add hooks for this function to all controllers that already have a
`SetResourceIdentifiers` hook.
`post_set_resource_identifiers` to