feature: Generate Code for resource adoption by annotation (#558)
Issue #, if available:

Description of changes:
These changes introduce a new generated function in all controllers 
that attempts to populate the resource spec/status with fields defined 
by the user. 

This change will require for developers to add hooks for this function to 
all controllers that already have a `SetResourceIdentifiers` hook.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.