Move `API` package from `aws-sdk-go` repo to this repository (#563)
Description of changes:
Since the aws-sdk-go repository will not be supported starting July 2025, 
we are moving it under our repository. Doing so will also make it easy to 
make any changes if needed, when we support aws-sdk-go-v2.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.