Renaming the converter.go function to LoadAPI, to hide the implementation details
Handling terminal errors and notFound errors differently.
In previous commit we were using the service defined exception structures.
This method does not work for controllers such as s3 and s3control. With more
tests, we will now be using the APIError type provided by aws/smithy-go. This interface is
implemented by all service errors so we will just compare to its ErrorCode() as we've done with V1.