Filter AWS injected tags
We are adding a new generated code that handles the filtering of tags.
These tags may be injected AWS resources during creation, or are added
to resources managed by cloudformation for example, and the controller
would see them in the diff and attempt to remove them. With these changes
we ensure that any tag that has `aws:` prefix will be filtered out after
the Find operation. Doing so will ensure that we don't attempt to remove
these tags.