Allow `from` `operation` to take output shapes
So currenly, when you define a `from` `operation` `path`
in generator.yaml fieldConfig, the code-generator checks for
the read_only configuration to decide whether to get the field
from the input shape or the output shape.

The change introduced here tries to retrieve spec fields from input
shapes first, if not found it checks output shape.
If read_only is configured to true, it does so in reverse order
(checks output shape first, if not found, checks input shape)